requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.16
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
pillow==11.0.0
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
rpds-py==0.22.1
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
